requirements.txt
# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# AI/LLM
google-generativeai>=0.3.0

# Web UI
streamlit>=1.28.0
plotly>=5.17.0
matplotlib>=3.10.0

# Utilities
python-dotenv>=1.0.0
